wip: implements wireframe pass

+++ source/Age.Rendering/Resources/VertexBuffer.cs
@@ -3,7 +3,7 @@ namespace Age.Rendering.Resources;
 public record VertexBuffer : Disposable
 {
     public required Buffer Buffer { get; init; }
-    public required ulong  Size   { get; init; }
+    public required uint   Size   { get; init; }
 
     protected override void OnDispose() =>
         this.Buffer.Dispose();
